+++ sprint2-andres/reserva_interfaz.py
@@ -53,6 +53,15 @@ class ReservaInterfaz(QWidget):
         self.nombre = self.input_nombre.text()
         self.tipo = self.combo_habitacion.currentText()
 
+        for i in self.nombre:
+            if i in '0123456789':
+                return QMessageBox.warning(self, "Error", "Por favor, ingrese un dato valido.") and self.input_nombre.clear()
+        
+        for i in self.nombre:
+            if i in '!"#$%&/()=?¡¿@,;.:-{[]^}<>¨´*+~':
+                return QMessageBox.warning(self, "Error", "Por favor, ingrese un dato valido.") and self.input_nombre.clear()
+
+
         
         if self.nombre and self.tipo:
             mensaje = f"Reserva Realizada:\n\nNombre: {self.nombre}\nHabitación: {self.tipo}"
@@ -78,6 +87,15 @@ class ReservaInterfaz(QWidget):
         archivo = 'sprint2-andres\\archivo_reservas.csv'
         self.nombre = self.input_nombre.text()
         self.tipo = self.combo_habitacion.currentText()
+
+        for i in self.nombre:
+            if i in '0123456789':
+                return QMessageBox.warning(self, "Error", "Por favor, ingrese un dato valido.") and self.input_nombre.clear()
+        
+        for i in self.nombre:
+            if i in '!"#$%&/()=?¡¿@,;.:-{[]^}<>¨´*+~':
+                return QMessageBox.warning(self, "Error", "Por favor, ingrese un dato valido.") and self.input_nombre.clear()
+
         
 
         if self.nombre and self.tipo:
@@ -95,14 +113,19 @@ class ReservaInterfaz(QWidget):
         
     
     def eliminar_reserva(self):
-        QMessageBox.information(self, "Reserva Eliminada", "Reserva eliminada exitosamente.")
         
         archivo = 'sprint2-andres\\archivo_reservas.csv'
         nombre_reserva = self.input_nombre.text()
         tipo = self.combo_habitacion.currentText()
 
         linea_reserva = self.encontrar_elemento_archivo(archivo, nombre_reserva, tipo)
-        self.eliminar_linea_archivo(archivo, linea_reserva)
+
+        if linea_reserva == None:
+            QMessageBox.warning(self, "Error", "Esa persona no tiene una reserva con esas características.")
+   
+        else:
+            self.eliminar_linea_archivo(archivo, linea_reserva)
+            QMessageBox.information(self, "Reserva Eliminada", "Reserva eliminada exitosamente.")
 
         self.mi_signal.emit()
